docs: arrival, department 파일 경로 수정
