+++ src/pages/booking/departLocation/index.tsx
@@ -1,0 +1,89 @@
+import { useState } from "react";
+import { Container, SearchBox, LocationList, LocationItem, LocationName, InfoIcon, RegionColumn, TerminalColumn } from "./index.styles";
+import TopBar from "../../../common/components/TopBar";
+import Input from "../../../common/components/Input";
+import terminalData from "../../../constants/terminal.json";
+import terminalRegionData from "../../../constants/terminal_region.json";
+
+import { useNavigate } from "@tanstack/react-router";
+
+import useSearchQueryStore from "../../../stores/useSearchQueryStore";
+
+interface Terminal {
+  tmnCd: number | string;
+  tmnNm: string;
+}
+
+interface TerminalRegion {
+  num: string;
+  tmnNm: string;
+}
+
+function LocationSelectPage() {
+  const { setSearchQuery } = useSearchQueryStore();
+  const navigate = useNavigate();
+  const [searchKeyword, setSearchKeyword] = useState("");
+
+  const terminals: Terminal[] = terminalData.response.body.items.item;
+  const terminalRegion: TerminalRegion[] = Object.entries(terminalRegionData)
+    .sort(([a], [b]) => a.localeCompare(b))
+    .map(([code, name]) => ({
+      num: code,
+      tmnNm: name
+    }));
+  
+  const filteredTerminals = terminals.filter(terminal => 
+    terminal.tmnNm.includes(searchKeyword)
+  )
+  
+  const handleTerminalClick = (terminal: Terminal) => {
+    setSearchQuery({
+      destId: terminal.tmnCd.toString(),
+    });
+    navigate({ to: '/booking'});
+    console.log(terminal.tmnCd);
+  }
+
+  return (
+    <div css={Container}>
+      <TopBar
+        leftSlot={null}
+        centerSlot={<h1>출발지를 선택하세요</h1>}
+        rightSlot={null}
+      />
+      
+      <div css={SearchBox}>
+        <Input
+          type="text"
+          placeholder="터미널, 지역명"
+          value={searchKeyword}
+          onValueChange={setSearchKeyword}
+        />
+      </div>
+
+      <div css={LocationList}>
+        <div css={RegionColumn}>
+        {terminalRegion.map((terminal) => (
+            <div css ={LocationItem} key={terminal.num}>
+              <div css ={LocationName}>{terminal.tmnNm}</div>
+            </div>
+          ))}
+      </div>
+
+      <div css={TerminalColumn}>
+        {filteredTerminals.map((terminal) => (
+            <div css ={LocationItem}
+             key={terminal.tmnCd}
+             onClick={() => handleTerminalClick(terminal)}
+            >
+              <div css ={LocationName}>{terminal.tmnNm}</div>
+              <div css ={InfoIcon}>i</div>
+            </div>
+          ))}
+      </div>
+      </div>
+    </div>
+  );
+}
+
+export default LocationSelectPage;
